Introduce Param object
Param object can be used to pass explicitly typed parameters to queries.
It is replacing Column object which is mainly intended to represent column information and therefore did not fit well for representing parameters.
But main reason is that flags in parameters have different meaning in columns.  E.g. fNullable in Column is set to 1 and represents nullable column, while flag 1 in parameters represents an output parameter.

Also changed default for Column objects to allow nullable values.