Implement token-based authentication (#182)
* Prelogin and Login handling for TDS74 FederatedAuth with Security Token

* only send fedauthrequired when token provided

* added unit tests for fedauth in prelogin and login parsing methods

* added access token login generation unit test

* type check update to pass new ruff check

* Update prelogin generation to enable nonce, correct nonce format in login7, move feature extension to match auth_packet offset

* corrected reading of featureextack, increased coverage

* print hex of nonce in log rather than raw bytes

* remove breakpoint from debug

* Validate access token and encode it as utf-8

* Add tests for access token validations

* Adjust cryptography version

This makes it installable on macOS without requiring a compiler.

* Remove nonce from connect and prelogin

According to the TDS spec, the nonce is sent by the server
initially, not by the client.

I also cross-checked with go-mssqldb which also doesn't take
a nonce from the user.

* Handle failing test

* Remove nonce from prelogin test

* Prevent usage of incompatible auth parameters

We raise an error when both user/password and access_token are
passed to `connect`.

* Add test for token-based auth

* Integrate fedauth test with connection.json

Since TENANT_ID and related settings are only used to test the
token-based connection, they can live side-by-side with the password-
auth settings.

* Implement access_token_callable

* Add example of usage to the readme

* Don't skip test if TENANT_ID is not provided

Since the test needs to run if HOST is set, it should error
if TENANT_ID is not provided.

* Rename "length" to "tokenlen"

* Encode token as UTF-16LE

* Move fedauth packet creation to separate function

* Document magic number

It's the size of the field that contains the token length.

* Raise on missing FEDAUTH from FEATUREEXTACK

This is when fedauth is used and specified in prelogin.

* Move access_token_callable to _connect

So that it can be used in the connection pool.

* Make sure a type of credential is provided

* Ignore again test if credentials are not set

* Try to exclude test from code coverage

* Add explicit check for access_token

This satisfies mypy.

* Remove incorrect check

It didn't account for SSO.

* Separate auth packet from extension packet

---------

Co-authored-by: Ian Bowler <ian.bowler@siemens-healthineers.com>